renamed helloworld, added some monitoring
