+++ database/spanner.go
@@ -8,6 +8,7 @@ import (
 
 	"cloud.google.com/go/spanner"
 	database "cloud.google.com/go/spanner/admin/database/apiv1"
+	"github.com/Cidan/sheep/stats"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/viper"
 	adminpb "google.golang.org/genproto/googleapis/spanner/admin/database/v1"
@@ -100,8 +101,12 @@ func (s *Spanner) Read(msg *Message) error {
 
 func (s *Spanner) Save(message *Message) error {
 	ctx := context.WithValue(context.Background(), contextKey("message"), message)
-	_, err := s.client.ReadWriteTransaction(ctx, s.doSave)
-	return err
+	if _, err := s.client.ReadWriteTransaction(ctx, s.doSave); err != nil {
+		stats.Incr("spanner.save.error", 1)
+		return err
+	}
+	stats.Incr("spanner.save.success", 1)
+	return nil
 }
 
 // Here's where the magic happens. Save out message!
